Initializable Cadence clients
